refactor youtube video duration parsing into a new file
