+++ server/tests/unit/services/youtube.spec.ts
@@ -533,20 +533,6 @@ describe("Youtube", () => {
 		});
 	});
 
-	describe("parseVideoLength", () => {
-		it.each([
-			["PT10S", 10],
-			["PT5M", 5 * 60],
-			["PT40M25S", 40 * 60 + 25],
-			["PT1H", 1 * 60 * 60],
-			["PT1H40M25S", 1 * 60 * 60 + 40 * 60 + 25],
-			["P1DT3S", 86403],
-			["P1D", 86400],
-		])("should parse %s into %s seconds", (timecode, seconds) => {
-			expect(adapter.parseVideoLength(timecode)).toEqual(seconds);
-		});
-	});
-
 	describe("getChannelId", () => {
 		it.each([
 			["https://youtube.com/@rollthedyc3", { handle: "@rollthedyc3" }],
